All code set up to dynamically perform training
