+++ Example/Podfile
@@ -4,6 +4,7 @@ platform :ios, '13.0'
 
 target 'Artificien_Example' do
   pod 'OpenMinedSwiftSyft', '~> 0.1.3-beta2'
+  pod 'Alamofire', '~> 4.7'
   pod 'Artificien', :path => '../'
 
 #  target 'Artificien_Tests' do
